gcc 9.4 compilation fixed
- copyright text added to sql_* headers;
- debugger start command added to micro-profiler_frontend project.

+++ frontend/src/untested/tables_ui.cpp
@@ -148,14 +148,14 @@ namespace micro_profiler
 		if (_initialized && hierarchical == _hierarchical && thread_id == _thread_id)
 			return;
 		if (hierarchical)
-			switch (thread_id)
+			switch (static_cast<int>(thread_id))
 			{
 			case threads_model::all: attach(representation<true, threads_all>::create(statistics_)); break;
 			case threads_model::cumulative: attach(representation<true, threads_cumulative>::create(statistics_)); break;
 			default: attach(representation<true, threads_filtered>::create(statistics_, thread_id)); break;
 			}
 		else
-			switch (thread_id)
+			switch (static_cast<int>(thread_id))
 			{
 			case threads_model::all: attach(representation<false, threads_all>::create(statistics_)); break;
 			case threads_model::cumulative: attach(representation<false, threads_cumulative>::create(statistics_)); break;
